change directory structure & update paths
