+++ src/core/App/Navigation/Header/index.js
@@ -1,6 +1,6 @@
 import { Logo, LogoWrapper, StyledLink, Title } from "./styled";
-import logo from "../../assets/logo.svg";
-import { toMovies } from "../../common/routes";
+import logo from "../../../../assets/logo.svg";
+import { toMovies } from "../../../../common/routes";
 
 export const Header = () => {
   return (
