+++ src/core/App/Navigation/index.js
@@ -4,7 +4,7 @@ import {
   NavigationWrapper,
   HeaderWrapper,
 } from "./styled";
-import { Search } from "../features/Search";
+import { Search } from "../../../features/Search";
 import { Header } from "./Header";
 import { Link } from "./Link";
