+++ src/core/App/index.js
@@ -2,19 +2,19 @@ import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { HashRouter, Redirect, Route, Switch } from "react-router-dom";
 import { Navigation } from "./Navigation";
-import MoviesPage from "./features/MoviesPage";
-import MoviePage from "./features/MoviePage";
-import PeoplePage from "./features/PeoplePage";
-import ProfilePage from "./features/ProfilePage";
-import { Error } from "./common/Error";
+import MoviesPage from "../../features/MoviesPage";
+import MoviePage from "../../features/MoviePage";
+import PeoplePage from "../../features/PeoplePage";
+import ProfilePage from "../../features/ProfilePage";
+import { Error } from "../../common/Error";
 import {
   toError,
   toMovie,
   toMovies,
   toPeople,
   toProfile,
-} from "./common/routes";
-import { fetchCommon } from "./common/commonSlice";
+} from "../../common/routes";
+import { fetchCommon } from "../../common/commonSlice";
 
 export const App = () => {
   const dispatch = useDispatch();
