+++ src/core/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./rootSaga";
-import searchReducer from "./features/Search/SearchSlice";
-import listReducer from "./features/listSlice";
-import commonReducer from "./common/commonSlice";
-import itemReducer from "./features/itemSlice";
+import searchReducer from "../features/Search/SearchSlice";
+import listReducer from "../features/listSlice";
+import commonReducer from "../common/commonSlice";
+import itemReducer from "../features/itemSlice";
 
 const sagaMiddleware = createSagaMiddleware();
