+++ src/features/listSaga.js
@@ -1,5 +1,5 @@
 import { call, takeLatest, put, delay } from "redux-saga/effects";
-import { getApiData } from "../getApiData";
+import { getApiData } from "../core/App/getApiData";
 import { fetchList, fetchListError, fetchListSuccess } from "./listSlice";
 import { apiKey, apiBase, language } from "./../common/commonValues";
